Move sort and offset and limit into Reasoner
implement SortedIterator to wrap non-Comparables

Revert changes to Iterators
