+++ common/iterator/sorted/SortedIterators.java
@@ -22,6 +22,7 @@ import com.vaticle.typedb.core.common.iterator.FunctionalIterator;
 import com.vaticle.typedb.core.common.iterator.Iterators;
 
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.List;
 import java.util.NavigableSet;
 import java.util.TreeSet;
@@ -58,11 +59,11 @@ public class SortedIterators {
 
     @SafeVarargs
     public static <T extends Comparable<? super T>, ORDER extends SortedIterator.Order> SortedIterator<T, ORDER> merge(SortedIterator<T, ORDER> iterator, SortedIterator<T, ORDER>... iterators) {
-        return new MergeMappedIterator<>(Iterators.iterate(list(list(iterators), iterator)), e -> e, iterator.order());
+        return new MergeMappedSortedIterator<>(Iterators.iterate(list(list(iterators), iterator)), e -> e, iterator.order());
     }
 
     public static <T extends Comparable<? super T>, ORDER extends SortedIterator.Order> SortedIterator<T, ORDER> merge(ORDER order, FunctionalIterator<SortedIterator<T, ORDER>> iterators) {
-        return new MergeMappedIterator<>(iterators, e -> e, order);
+        return new MergeMappedSortedIterator<>(iterators, e -> e, order);
     }
 
     public static <T extends Comparable<? super T>, ORDER extends SortedIterator.Order> SortedIterator<T, ORDER> onConsume(SortedIterator<T, ORDER> iterator, Runnable onConsume) {
@@ -109,11 +110,11 @@ public class SortedIterators {
 
         @SafeVarargs
         public static <T extends Comparable<? super T>, ORDER extends SortedIterator.Order> SortedIterator.Forwardable<T, ORDER> merge(SortedIterator.Forwardable<T, ORDER> iterator, SortedIterator.Forwardable<T, ORDER>... iterators) {
-            return new MergeMappedIterator.Forwardable<>(Iterators.iterate(list(list(iterators), iterator)), e -> e, iterator.order());
+            return new MergeMappedSortedIterator.Forwardable<>(Iterators.iterate(list(list(iterators), iterator)), e -> e, iterator.order());
         }
 
         public static <T extends Comparable<? super T>, ORDER extends SortedIterator.Order> SortedIterator.Forwardable<T, ORDER> merge(FunctionalIterator<SortedIterator.Forwardable<T, ORDER>> iterators, ORDER order) {
-            return new MergeMappedIterator.Forwardable<>(iterators, e -> e, order);
+            return new MergeMappedSortedIterator.Forwardable<>(iterators, e -> e, order);
         }
 
         @SafeVarargs
