+++ reasoner/Reasoner.java
@@ -28,6 +28,7 @@ import com.vaticle.typedb.core.common.parameters.Label;
 import com.vaticle.typedb.core.common.parameters.Options;
 import com.vaticle.typedb.core.concept.ConceptManager;
 import com.vaticle.typedb.core.concept.answer.ConceptMap;
+import com.vaticle.typedb.core.concept.thing.Attribute;
 import com.vaticle.typedb.core.concept.thing.Thing;
 import com.vaticle.typedb.core.concept.type.Type;
 import com.vaticle.typedb.core.concurrent.producer.Producer;
@@ -42,17 +43,27 @@ import com.vaticle.typedb.core.reasoner.answer.Explanation;
 import com.vaticle.typedb.core.reasoner.controller.ControllerRegistry;
 import com.vaticle.typedb.core.traversal.TraversalEngine;
 import com.vaticle.typedb.core.traversal.common.Identifier;
+import com.vaticle.typeql.lang.common.TypeQLArg;
+import com.vaticle.typeql.lang.pattern.variable.Reference;
 import com.vaticle.typeql.lang.pattern.variable.UnboundVariable;
 import com.vaticle.typeql.lang.query.TypeQLMatch;
 
+import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
+import java.util.stream.IntStream;
 
 import static com.vaticle.typedb.common.collection.Collections.list;
 import static com.vaticle.typedb.common.collection.Collections.set;
+import static com.vaticle.typedb.core.common.exception.ErrorMessage.Internal.ILLEGAL_STATE;
 import static com.vaticle.typedb.core.common.exception.ErrorMessage.Pattern.UNSATISFIABLE_PATTERN;
 import static com.vaticle.typedb.core.common.exception.ErrorMessage.Pattern.UNSATISFIABLE_SUB_PATTERN;
+import static com.vaticle.typedb.core.common.exception.ErrorMessage.ThingRead.INVALID_THING_CASTING;
+import static com.vaticle.typedb.core.common.exception.ErrorMessage.ThingRead.SORT_ATTRIBUTE_NOT_COMPARABLE;
+import static com.vaticle.typedb.core.common.exception.ErrorMessage.ThingRead.SORT_VARIABLE_NOT_ATTRIBUTE;
 import static com.vaticle.typedb.core.common.iterator.Iterators.iterate;
 import static com.vaticle.typedb.core.common.parameters.Arguments.Query.Producer.EXHAUSTIVE;
 import static com.vaticle.typedb.core.concurrent.executor.Executors.PARALLELISATION_FACTOR;
@@ -73,9 +84,7 @@ public class Reasoner {
         this.conceptMgr = conceptMgr;
         this.traversalEng = traversalEng;
         this.logicMgr = logicMgr;
-        Context.Query defaultContext = new Context.Query(context, new Options.Query());
-        defaultContext.producer(Either.first(EXHAUSTIVE));
-        this.controllerRegistry = new ControllerRegistry(actor(), traversalEng, conceptMgr, logicMgr, defaultContext);
+        this.controllerRegistry = new ControllerRegistry(actor(), traversalEng, conceptMgr, logicMgr, context);
         this.explainablesManager = new ExplainablesManager();
     }
 
@@ -83,31 +92,16 @@ public class Reasoner {
         return controllerRegistry;
     }
 
-    private boolean mayReason(Disjunction disjunction, Context.Query context) {
-        if (!context.options().infer() || context.transactionType().isWrite() || !logicMgr.rules().hasNext()) {
-            return false;
-        }
-        return mayReason(disjunction);
-    }
-
-    private boolean mayReason(Disjunction disjunction) {
-        for (Conjunction conj : disjunction.conjunctions()) {
-            Set<Variable> vars = conj.variables();
-            List<Negation> negs = conj.negations();
-            if (iterate(vars).flatMap(v -> iterate(v.inferredTypes())).distinct().anyMatch(this::hasRule)) return true;
-            if (!negs.isEmpty() && iterate(negs).anyMatch(n -> mayReason(n.disjunction()))) return true;
-        }
-        return false;
-    }
-
-    private boolean hasRule(Label type) {
-        return logicMgr.rulesConcluding(type).hasNext() || logicMgr.rulesConcludingHas(type).hasNext();
-    }
-
     public FunctionalIterator<ConceptMap> execute(Disjunction disjunction, TypeQLMatch.Modifiers modifiers, Context.Query context) {
         inferAndValidateTypes(disjunction);
-        if (mayReason(disjunction, context)) return executeReasoner(disjunction, filter(modifiers.filter()), context);
-        else return executeTraversal(disjunction, context, filter(modifiers.filter()));
+        FunctionalIterator<ConceptMap> answers;
+        if (mayReason(disjunction, context)) answers = executeReasoner(disjunction, filter(modifiers.filter()), context);
+        else answers = executeTraversal(disjunction, context, filter(modifiers.filter()));
+
+        if (modifiers.sort().isPresent()) answers = sort(answers, modifiers.sort().get());
+        if (modifiers.offset().isPresent()) answers = answers.offset(modifiers.offset().get());
+        if (modifiers.limit().isPresent()) answers = answers.limit(modifiers.limit().get());
+        return answers;
     }
 
     private void inferAndValidateTypes(Disjunction disjunction) {
@@ -122,12 +116,6 @@ public class Reasoner {
         }
     }
 
-    private Set<Identifier.Variable.Retrievable> filter(List<UnboundVariable> typeQLVars) {
-        Set<Identifier.Variable.Retrievable> names = new HashSet<>();
-        iterate(typeQLVars).map(v -> v.reference().asName()).map(Identifier.Variable::of).forEachRemaining(names::add);
-        return names;
-    }
-
     private Set<Conjunction> incoherentConjunctions(Disjunction disjunction) {
         assert !disjunction.isCoherent();
         Set<Conjunction> causes = new HashSet<>();
@@ -138,6 +126,87 @@ public class Reasoner {
             else incoherentNegs.forEachRemaining(n -> causes.addAll(incoherentConjunctions(n.disjunction())));
         }
         return causes;
+    }
+
+    private boolean mayReason(Disjunction disjunction, Context.Query context) {
+        if (!context.options().infer() || context.transactionType().isWrite() || !logicMgr.rules().hasNext()) {
+            return false;
+        }
+        return mayReason(disjunction);
+    }
+
+    private boolean mayReason(Disjunction disjunction) {
+        for (Conjunction conj : disjunction.conjunctions()) {
+            Set<Variable> vars = conj.variables();
+            List<Negation> negs = conj.negations();
+            if (iterate(vars).flatMap(v -> iterate(v.inferredTypes())).distinct().anyMatch(this::ruleConcludes)) {
+                return true;
+            }
+            if (!negs.isEmpty() && iterate(negs).anyMatch(n -> mayReason(n.disjunction()))) return true;
+        }
+        return false;
+    }
+
+    private boolean ruleConcludes(Label type) {
+        return logicMgr.rulesConcluding(type).hasNext() || logicMgr.rulesConcludingHas(type).hasNext();
+    }
+
+    private Set<Identifier.Variable.Retrievable> filter(List<UnboundVariable> typeQLVars) {
+        Set<Identifier.Variable.Retrievable> names = new HashSet<>();
+        iterate(typeQLVars).map(v -> v.reference().asName()).map(Identifier.Variable::of).forEachRemaining(names::add);
+        return names;
+    }
+
+    private FunctionalIterator<ConceptMap> sort(FunctionalIterator<ConceptMap> answers, Sortable.Sorting sorting) {
+        List<Reference.Name> sortVars = iterate(sorting.vars()).map(var -> var.reference().asName()).toList();
+        Comparator<List<Attribute>> multiComparator = multiComparator(sortVars.size());
+        Comparator<ConceptMap> comparator = (answer1, answer2) -> {
+            List<Attribute> attributes1 = new ArrayList<>(sortVars.size());
+            List<Attribute> attributes2 = new ArrayList<>(sortVars.size());
+            for (Reference.Name var : sortVars) {
+                try {
+                    Attribute att1 = answer1.get(var).asAttribute();
+                    Attribute att2 = answer2.get(var).asAttribute();
+                    if (!att1.getType().getValueType().comparables().contains(att2.getType().getValueType())) {
+                        throw TypeDBException.of(SORT_ATTRIBUTE_NOT_COMPARABLE, var);
+                    }
+                    attributes1.add(att1);
+                    attributes2.add(att2);
+                } catch (TypeDBException e) {
+                    if (e.code().isPresent() || e.code().get().equals(INVALID_THING_CASTING.code())) {
+                        throw TypeDBException.of(SORT_VARIABLE_NOT_ATTRIBUTE, var);
+                    } else {
+                        throw e;
+                    }
+                }
+            }
+            return multiComparator.compare(attributes1, attributes2);
+        };
+        comparator = (sorting.order() == TypeQLArg.Order.DESC) ? comparator.reversed() : comparator;
+        return iterate(answers.stream().sorted(comparator).iterator());
+    }
+
+    private Comparator<List<Attribute>> multiComparator(int n) {
+        Optional<Comparator<List<Attribute>>> multiComparator = IntStream.range(0, n)
+                .mapToObj(i -> Comparator.comparing((List<Attribute> attrs) -> attrs.get(i), (att1, att2) -> {
+                    if (att1.isString()) {
+                        return att1.asString().getValue().compareToIgnoreCase(att2.asString().getValue());
+                    } else if (att1.isBoolean()) {
+                        return att1.asBoolean().getValue().compareTo(att2.asBoolean().getValue());
+                    } else if (att1.isLong() && att2.isLong()) {
+                        return att1.asLong().getValue().compareTo(att2.asLong().getValue());
+                    } else if (att1.isDouble() || att2.isDouble()) {
+                        Double double1 = att1.isLong() ? att1.asLong().getValue() : att1.asDouble().getValue();
+                        Double double2 = att2.isLong() ? att2.asLong().getValue() : att2.asDouble().getValue();
+                        return double1.compareTo(double2);
+                    } else if (att1.isDateTime()) {
+                        return (att1.asDateTime().getValue()).compareTo(att2.asDateTime().getValue());
+                    } else {
+                        throw TypeDBException.of(ILLEGAL_STATE);
+                    }
+                })).reduce(Comparator::thenComparing);
+        assert multiComparator.isPresent();
+        return multiComparator.get();
     }
 
     public FunctionalIterator<ConceptMap> executeReasoner(Disjunction disjunction, Set<Identifier.Variable.Retrievable> filter,
